+++ sqllin-driver/src/nativeMain/kotlin/com/ctrip/sqllin/driver/SQLiteStatement.kt
@@ -23,13 +23,15 @@ package com.ctrip.sqllin.driver
 
 internal interface SQLiteStatement {
 
+    fun isNull(columnIndex: Int): Boolean
+
     fun columnGetLong(columnIndex: Int): Long
 
     fun columnGetDouble(columnIndex: Int): Double
 
-    fun columnGetString(columnIndex: Int): String
+    fun columnGetString(columnIndex: Int): String?
 
-    fun columnGetBlob(columnIndex: Int): ByteArray
+    fun columnGetBlob(columnIndex: Int): ByteArray?
 
     fun columnCount(): Int